// src/app/interfaces/book.interface.ts
export interface IBook {
  title: string;
  author: string;
  genre:
    | "FICTION"
    | "NON_FICTION"
    | "SCIENCE"
    | "HISTORY"
    | "BIOGRAPHY"
    | "FANTASY";
  isbn: string;
  description?: string;
  copies: number;
  available: boolean;
}

// instance method
export interface BookMethod {
  updateBookAfterBorrow?(quantity: number): Promise<IBook>;
}